refactor: centralizando beans de email em MailSenderConfig
