+++ src/main/java/com/ufrn/nei/almoxarifadoapi/config/MailSenderConfig.java
@@ -1,19 +1,46 @@
 package com.ufrn.nei.almoxarifadoapi.config;
 
+import java.util.Properties;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.mail.SimpleMailMessage;
+
+import jakarta.mail.MessagingException;
+import jakarta.mail.Session;
+import jakarta.mail.internet.InternetAddress;
+import jakarta.mail.internet.MimeMessage;
 
 @Configuration
 public class MailSenderConfig {
     @Value("${spring.mail.username}")
     private String sender;
 
+    @Value("${spring.mail.host}")
+    private String host;
+
+    @Value("${spring.mail.port}")
+    private String port;
+
     @Bean
     SimpleMailMessage configSimpleMailMessage() {
         SimpleMailMessage simpleMailMessage = new SimpleMailMessage();
         simpleMailMessage.setFrom(sender);
         return simpleMailMessage;
+    }
+
+    @Bean
+    MimeMessage configMimeMessage() throws MessagingException {
+        Properties props = new Properties();
+        props.put("mail.smtp.host", host);
+        props.put("mail.smtp.port", port);
+        props.put("mail.smtp.user", sender);
+
+        Session session = Session.getInstance(props);
+
+        MimeMessage message = new MimeMessage(session);
+        message.setFrom(new InternetAddress(sender));
+        return message;
     }
 }
